Partially complete models no longer added early
As of the last major update to Promenade, a bug was introduced that
caused partially complete models to be "added" early to their respective
collections. We didn't have a test that checked the completeness of
models at the point at which `add` events were triggered, so this got
past us. The problem should be resolved with this fix.